+++ environment.sh
@@ -11,7 +11,7 @@
 export BOOT2DOCKER_EXTENSION_DIR="/var/lib/boot2docker/extension"
 export BOOT2DOCKER_EXTENSION_LIB_DIR="$BOOT2DOCKER_EXTENSION_DIR/lib"
 export BOOT2DOCKER_EXTENSION_PARAM_DIR="${BOOT2DOCKER_EXTENSION_DIR}/param.d"
-export BOOT2DOCKER_DK_IMAGES_SAVE_DIR="/vagrant/.vagrant/b2d_dk_images_backup"
+export BOOT2DOCKER_DK_IMAGES_SAVE_DIR="/vagrant/.vagrant/b2d_common/dk_images_backup"
 export HOME_DOCKER_USER="/home/docker"
 export CRON_DK_IPULL_LOGFILE="/var/log/cron_dk_ipull.log"
 export CRON_DK_IBACKUP_LOGFILE="/var/log/cron_dk_ibackup.log"