+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## 17.XX.X-ce (2017-08-09)
 - UPDATE: version check sleep moved from 5 sec to 3 sec to avoid so much wait
+- UPDATE: moving docker image backup to folder /vagrant/.vagrant/b2d_common/dk_images_backup for common share between various VM with last version of b2d vagrant box
 
 ## 17.06.0-ce (2017-07-05)
 - UPDATE: update to boot2docker 17.06.0-ce